refactor(day-diet): extract state management to dedicated store
