+++ src/routes/test-app.tsx
@@ -9,7 +9,7 @@ import {
 import {
   setTargetDay,
   targetDay,
-} from '~/modules/diet/day-diet/application/dayDiet'
+} from '~/modules/diet/day-diet/application/dayDietStore'
 import {
   createNewDayDiet,
   type DayDiet,
